Refactoring: Addressing Reviews
-Separated user_data.py UserConfig class away from constants
-Changing import for UserConfig from user_data rather than constants
-Changing Config to UserConfig, to reflect the class name change
